requirements.txt
aenum==2.2.4
boto3==1.14.27
botocore==1.17.27
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
docutils==0.15.2
idna==2.10
imageio==2.9.0
jmespath==0.10.0
joblib==0.16.0
kiwisolver==1.2.0
lightgbm==2.3.1
matplotlib==3.3.0
networkx==2.4
numpy==1.19.1
oauthlib==3.1.0
Pillow==7.2.0
pyparsing==2.4.7
pyproj==2.6.1.post1
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.24.0
requests-oauthlib==1.3.0
s2cloudless==1.4.0
s3transfer==0.3.3
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.5.2
sentinelhub==3.0.4
Shapely==1.7.0
six==1.15.0
threadpoolctl==2.1.0
tifffile==2020.7.22
urllib3==1.25.10
utm==0.5.0
